// crates/opsml_server/opsml_ui/src/lib/scripts/types.ts
export interface FileInfo {
  name: string;
  size: number;
  type: string;
  created: number;
  islink: boolean;
  mode: number;
  uid: number;
  gid: number;
  mtime: number;
  ino: number;
  nlink: number;
  uri: string;
  suffix: string;
}

export interface ViewContent {
  content?: string;
  view_type?: string;
}

export interface FileView {
  file_info: FileInfo;
  content: ViewContent;
}

export interface Files {
  files: FileInfo[];
  mtime: number;
}